It works :letsgo: Issue was that the mat3x3 was still messed up :pogfish:
Currently it looks like it scales the image too.
That is because the uv coordinates are from zero to 1.0, regardless of
how the dimensions are relative to each other. That means that it will
appear scaled. But since the real project will have the image always
square, this isn't actually an issue.